Rename Queue dimensions to Product Dimensions
